[Move #359] livekit API 폴더 생성 및 경로 변경
